// types.ts
export interface Language {
  code: string;
  name: string;
}

export type Translations = Record<string, string>;

export interface Example {
  sentence: string;
  translation: string;
}

export type Examples = Record<string, Example[]>;

export interface Card {
  id?: string; // ID is optional as it's generated by the DB
  frontText: string;
  targetLanguage: string;
  sourceLanguages: string[];
  translations: Translations;
  examples: Examples;
  notes?: string;
  created_at?: string;
}

export interface Settings {
  targetLanguage: string;
  sourceLanguages: string[];
}